Refactor code for deferred config updates
This commit refactors the code to allow for deferred updates to the configuration options without having to restart the game. It includes changing to a functional approach for fetching parent storage options. This implementation also corrects the workbench range usage and makes similar changes in other storage options classes. Versioning and the changelog have been updated accordingly.