Remove Definitions.cs and refactor related dependencies
The Definitions class was removed and its usage was replaced in the AssetHandler and GarbageCanManager classes. Additionally, the version of the project was updated. Several changes to how assets and garbage can management are handled were implemented. These changes included updating how garbage cans are found and invalidated, as well as changes to the event subscriptions in the GarbageCanManager.