Refactor container retrieval and add supporting features
The code was refactored for getting a container from buildings to streamline the process and improve efficiency. Support was added for Shipping Bins and the feature to show the chest being configured was introduced. Several bugs were fixed including config options reflecting without a game restart and an issue with Workbenches and Satchels not opening correctly. Update keys in the BetterChests.csproj were also modified.