Update BetterChests for FauxCore API changes
This commit includes updates to keep BetterChests compatible with recent changes in the FauxCore API. In AssetHandler.cs, a new dependency ('modContentHelper') has been added for accessing mod content, and has been utilized to load assets in 'themeHelper'. An unnecessary code block in the file was removed for efficiency. In the CHANGELOG.md, these changes are documented.