Update CustomBush to version 1.2.1 and add Id property
This commit updates the version of CustomBush from 1.2.0 to 1.2.1 and introduces an Id property in the ICustomBush interface. The id is injected into the custom bushes in the AssetHandler when the data is loaded, and it is stored inside the data model for all custom bushes. This allows better identification of custom bushes within the application.