Refactor container classes to simplify item menu handling
The commit comprises a significant refactor of various container classes in the BetterChests plugin. The update streamlined the handling of the item menu, reducing redundant and complex code. Updating the 'ShowMenu' methods has enabled a more straightforward approach to defining different containers' behaviors. The 'SourceItem' property has been removed as it's no longer necessary with the newer, more efficient organization of the code. This commit will make future maintenance and enhancements simpler to implement.