Update BetterChests features and fix storage display issue
This commit updates the features of the BetterChests mod, adding configuration options for new storage types and allowing more types of furniture to be configured. It also addresses an issue where chests were incorrectly displaying the inventory of adjacent chests. Additionally, the version number is updated according to these changes.